Refactor recalibration / recentering, clean up code, add basic keyboard handler
