+++ pyproject.toml
@@ -22,6 +22,7 @@ winotify = [
 onnxruntime = "^1.13.1"
 colorama = "^0.4.6"
 taskipy = "^1.10.4"
+pynput = "^1.7.6"
 [tool.poetry.group.dev.dependencies]
 black = "^22.10.0"
 pyinstaller = "^5.6.2"
